refactor(core): split candidate datum, layout chrome, and smooth geometry
Extract frame-row mapping and aggregate lineage filtering from candidate
datums; pull labs/formatters/legends into panel-layout chrome; split smooth
ribbon vs fitted line. Tests pin bin lineage closedness, legend margin, and
smooth SE ribbon batches.
