+++ packages/core/src/pipeline/geometry-smooth-line.ts
@@ -1,0 +1,59 @@
+/**
+ * Smooth fitted-line path batch.
+ */
+import type { SmoothParams } from "@ggsvelte/spec";
+
+import type { GeometryBatch } from "../scene.js";
+
+import type { LayerFrame, ResolvedColorScale } from "./types.js";
+import type { Frame } from "./geometry-shared.js";
+import { positionOf } from "./geometry-shared.js";
+import { DEFAULT_SMOOTH_LINEWIDTH, groupColor } from "./geometry-smooth-shared.js";
+
+export function buildSmoothLineBatch(input: {
+  frame: LayerFrame;
+  fx: Frame;
+  color: ResolvedColorScale | null;
+  groupRows: readonly (readonly number[])[];
+}): GeometryBatch {
+  const { frame, fx, color, groupRows } = input;
+  const { binding } = frame;
+  const params = (binding.layer.params ?? {}) as SmoothParams;
+
+  let total = 0;
+  for (const rows of groupRows) total += rows.length;
+  const positions = new Float32Array(total * 2);
+  const rowIndex = new Uint32Array(total);
+  const pathOffsets = new Uint32Array(groupRows.length + 1);
+  const strokes: (string | null)[] = [];
+  let cursor = 0;
+  for (let s = 0; s < groupRows.length; s++) {
+    pathOffsets[s] = cursor;
+    const rows = groupRows[s]!;
+    for (const row of rows) {
+      const tx = positionOf(fx.xScale, frame.xNumeric, frame.xValues, row);
+      const ty = positionOf(fx.yScale, frame.yNumeric, null, row);
+      positions[cursor * 2] = tx * fx.innerWidth;
+      positions[cursor * 2 + 1] = fx.innerHeight - ty * fx.innerHeight;
+      rowIndex[cursor] = frame.rowIndex[row]!;
+      cursor++;
+    }
+    strokes.push(
+      groupColor(color, frame.colorValues, binding.color.scaledConstant, rows[0]!) ??
+        binding.color.constant,
+    );
+  }
+  pathOffsets[groupRows.length] = cursor;
+  return {
+    kind: "paths",
+    layerIndex: binding.index,
+    panelIndex: 0,
+    positions,
+    rowIndex,
+    pathOffsets,
+    strokes,
+    linewidth: params.linewidth ?? DEFAULT_SMOOTH_LINEWIDTH,
+    alpha: params.alpha ?? 1,
+    curve: "linear",
+  };
+}
